separated file and model parsers for clarity
