+++ src/model/__init__.py
@@ -12,8 +12,5 @@ from .FileItem import FileItem
 from .Mission import Mission
 from .MissionCompiler import MissionCompiler
 from .MissionComponents import MissionComponents
-from .MissionFileParser import MissionFileParser
-from .MissionParser import MissionParser
-from .TriggerParser import TriggerParser
 
 import src.model.components
