+++ src/model/file_data_parsers/__init__.py
@@ -1,0 +1,10 @@
+"""
+file_data_parsers
+=================
+
+This package contains parsers for ES mission files
+"""
+
+
+from .FileMissionItemParser import FileMissionItemParser
+from .MissionFileParser import MissionFileParser
