+++ src/model/model_data_parsers/TriggerParser.py
