+++ src/model/model_data_parsers/__init__.py
@@ -1,0 +1,10 @@
+"""
+model_data_parsers
+==================
+
+This package contains parsers for ESMB model data
+"""
+
+
+from .MissionParser import MissionParser
+from .TriggerParser import TriggerParser
